gobject: move() hoist player sign-extend by widening to int
const s8 player re-emitted extsb. on every player==0 reuse; the target
sign-extends m_animStateMisc once and keeps player as a full int. Cast
to int via (s8) does the single extsb. move 90.56->91.07, unit
91.439->91.494.

+++ src/gobject.cpp
@@ -528,7 +528,7 @@ void CGObject::move()
 
         movingWithScript = true;
     } else {
-        const s8 player = m_animStateMisc;
+        const int player = static_cast<s8>(m_animStateMisc);
         if ((player >= 0)
             && (player < 4)
             && m_weaponNodeFlagAll.m_bits1.m_shield
